sql injection guard + permissions

+++ server/app/collections.php
@@ -4,7 +4,7 @@ class Collections {
   public function __construct($server) {
     $this->server = $server;
   }
-  
+
   function generate_join_clause($table) {
     $sql = '';
     $joinData = $this->server->moma->joinData[$table];
@@ -20,8 +20,12 @@ class Collections {
     return $sql;
   }
   
-  function generate_where_clause($constraints, $careful = false) {
+  function generate_where_clause($constraints, $careful = false, $modify = false) {
     $sql = '';
+    
+    if ($modify) {
+      $constraints['userid'] = $this->server->user['id'];
+    }
     
     if (!empty($constraints)) {
       $sql .= " WHERE ";
@@ -30,9 +34,14 @@ class Collections {
       foreach ($constraints as $column => $value) {
         if ($first) $sql .= " AND ";
         
-        if (is_string($value))
+        if (is_string($value)) {
+          $value = mysql_real_escape_string($value);
           $value = "'$value'";
-        
+        }
+
+        if (!ctype_alpha(str_replace(array('.', '_'), '', $column)))
+          throw new Error("");
+          
         $sql .= "$column = $value";
         $first = true;
       }
@@ -62,9 +71,15 @@ class Collections {
   
   function read($desc) {
     extract($desc);
-    
+        
     if ($properties == '*')
       $properties = $this->generate_all_properties($table);
+
+    if (!ctype_alpha(str_replace(array('*', '_', '.', ' ', ','), '', $properties)))
+      throw new Error("");
+    
+    if (!ctype_alpha($table))
+      throw new Error("");
     
     $sql = "SELECT $properties FROM $table";
     $sql .= $this->generate_join_clause($table);
@@ -78,11 +93,18 @@ class Collections {
       else
         throw new Error('orderby first value must be "<" or ">"');  
 
+      if (!ctype_alpha(str_replace(array('.', '_'), '', $orderby[1])))
+        throw new Error("");
+
       $sql .= " ORDER BY $orderby[1] $ascdesc"; 
     }
 
     if (!empty($range)) {
       extract($range);
+
+      if (!ctype_digit($count) || !ctype_digit($startIndex))
+        throw new Error("");
+
       $sql .= " LIMIT $count OFFSET $startIndex";
     }
 
@@ -105,11 +127,15 @@ class Collections {
       return $rows;
     }
   }
-
+  
   function delete($desc) {
     extract($desc);
+
+    if (!ctype_alpha($table))
+      throw new Error("");
+    
     $sql = "DELETE FROM $table";
-    $sql .= $this->generate_where_clause($constraints, true);
+    $sql .= $this->generate_where_clause($constraints, true, true);
     $query = $this->server->moma->query($sql);
     return $query->rowCount();    
   }
@@ -120,18 +146,24 @@ class Collections {
     
     $values['modified'] = time();
 
+    if (!ctype_alpha($table))
+      throw new Error("");
+    
     $sql = "UPDATE $table SET ";
    
     $valuesSql = array();
     foreach ($values as $key => $value) {
+      if (!ctype_alpha(str_replace(array('_', '.'), '', $key)))
+        throw new Error("");
+
       if (is_string($value))
-        $value = "'".sqlite_escape_string($value)."'";
+        $value = "'".mysql_escape_string($value)."'";
         
       $valuesSql[] = "$key = $value";
     }
                  
     $sql .= implode(', ', $valuesSql);                                             
-    $sql .= $this->generate_where_clause($constraints);
+    $sql .= $this->generate_where_clause($constraints, false, true);
     
     $query = $this->server->moma->query($sql);
     $result = $query->rowCount();
@@ -141,6 +173,10 @@ class Collections {
 
   function create($desc) {
     extract($desc);
+
+    if (!ctype_alpha($table))
+      throw new Error("");
+    
     $sql = "INSERT INTO $table ";
 
     $valuesSql = array();
@@ -150,6 +186,9 @@ class Collections {
 
     $columns = implode(', ', array_keys($values));
     foreach ($values as $key => $value) {
+      if (!ctype_alpha(str_replace(array('_', '.'), '', $key)))
+        throw new Error("");
+
       if (is_string($value))
         $value = "'".sqlite_escape_string($value)."'";
